Extract json file IO to a separate class
Extremely basic right now with just the read/write funcs but can be expanded later
